requirements.txt
fastapi
uvicorn
pydantic-settings
python-dotenv
pyyaml
loguru
sqlalchemy
asyncpg
aiosqlite
greenlet
openai
redis
hiredis
jsonschema
pyjwt
pytest
pytest-asyncio
httpx
pytest-env
